refactor: uikit tabbar
