+++ Feather/Entry.swift
@@ -1,0 +1,216 @@
+//
+//  FeatherApp.swift
+//  Feather
+//
+//  Created by samara on 10.04.2025.
+//
+
+import SwiftUI
+import Nuke
+import IDeviceSwift
+import OSLog
+
+@main enum Entry {
+	static func main() {
+		let delegate = AppDelegate()
+		UIApplication.shared.delegate = delegate
+		_ = UIApplicationMain(CommandLine.argc, CommandLine.unsafeArgv, nil, NSStringFromClass(AppDelegate.self))
+	}
+}
+
+final class AppDelegate: UIResponder, UIApplicationDelegate {
+	var window: UIWindow?
+	
+	func application(
+		_ application: UIApplication,
+		didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
+	) -> Bool {
+		_createPipeline()
+		_createDocumentsDirectories()
+		ResetView.clearWorkCache()
+		_addDefaultCertificates()
+		
+		let tc = TabController()
+		window = UIWindow(frame: UIScreen.main.bounds)
+		window?.rootViewController = tc
+		window?.makeKeyAndVisible()
+		
+		DispatchQueue.main.async {
+			self.window!.tintColor = UIColor(Color(hex: UserDefaults.standard.string(forKey: "Feather.userTintColor") ?? "#848ef9"))
+			self.window!.overrideUserInterfaceStyle = UIUserInterfaceStyle(rawValue: UserDefaults.standard.integer(forKey: "Feather.userInterfaceStyle")) ?? .unspecified
+		}
+		
+		return true
+	}
+	
+	private func _createPipeline() {
+		DataLoader.sharedUrlCache.diskCapacity = 0
+		
+		let pipeline = ImagePipeline {
+			let dataLoader: DataLoader = {
+				let config = URLSessionConfiguration.default
+				config.urlCache = nil
+				return DataLoader(configuration: config)
+			}()
+			
+			let dataCache = try? DataCache(name: "\(Bundle.main.bundleIdentifier!).datacache")
+			let imageCache = Nuke.ImageCache()
+			
+			dataCache?.sizeLimit = 500 * 1024 * 1024
+			imageCache.costLimit = 100 * 1024 * 1024
+			
+			$0.dataCache = dataCache
+			$0.imageCache = imageCache
+			$0.dataLoader = dataLoader
+			$0.dataCachePolicy = .automatic
+			$0.isStoringPreviewsInMemoryCache = false
+		}
+		
+		ImagePipeline.shared = pipeline
+	}
+	
+	private func _createDocumentsDirectories() {
+		let fileManager = FileManager.default
+
+		let directories: [URL] = [
+			fileManager.archives,
+			fileManager.certificates,
+			fileManager.signed,
+			fileManager.unsigned
+		]
+		
+		for url in directories {
+			try? fileManager.createDirectoryIfNeeded(at: url)
+		}
+	}
+	
+	private func _addDefaultCertificates() {
+		guard
+			UserDefaults.standard.bool(forKey: "feather.didImportDefaultCertificates") == false,
+			let signingAssetsURL = Bundle.main.url(forResource: "signing-assets", withExtension: nil)
+		else {
+			return
+		}
+		
+		do {
+			let folderContents = try FileManager.default.contentsOfDirectory(
+				at: signingAssetsURL,
+				includingPropertiesForKeys: nil,
+				options: .skipsHiddenFiles
+			)
+			
+			for folderURL in folderContents {
+				guard folderURL.hasDirectoryPath else { continue }
+				
+				let certName = folderURL.lastPathComponent
+				
+				let p12Url = folderURL.appendingPathComponent("cert.p12")
+				let provisionUrl = folderURL.appendingPathComponent("cert.mobileprovision")
+				let passwordUrl = folderURL.appendingPathComponent("cert.txt")
+				
+				guard
+					FileManager.default.fileExists(atPath: p12Url.path),
+					FileManager.default.fileExists(atPath: provisionUrl.path),
+					FileManager.default.fileExists(atPath: passwordUrl.path)
+				else {
+					Logger.misc.warning("Skipping \(certName): missing required files")
+					continue
+				}
+				
+				let password = try String(contentsOf: passwordUrl, encoding: .utf8)
+				
+				FR.handleCertificateFiles(
+					p12URL: p12Url,
+					provisionURL: provisionUrl,
+					p12Password: password,
+					certificateName: certName,
+					isDefault: true
+				) { _ in
+					
+				}
+			}
+			UserDefaults.standard.set(true, forKey: "feather.didImportDefaultCertificates")
+		} catch {
+			Logger.misc.error("Failed to list signing-assets: \(error)")
+		}
+	}
+	
+	func application(_: UIApplication, open url: URL, options _: [UIApplication.OpenURLOptionsKey: Any] = [:]) -> Bool {
+		if url.scheme == "feather" {
+			/// feather://import-certificate?p12=<base64>&mobileprovision=<base64>&password=<base64>
+			if url.host == "import-certificate" {
+				guard
+					let components = URLComponents(url: url, resolvingAgainstBaseURL: false),
+					let queryItems = components.queryItems
+				else {
+					return false
+				}
+				
+				func queryValue(_ name: String) -> String? {
+					queryItems.first(where: { $0.name == name })?.value?.removingPercentEncoding
+				}
+				
+				guard
+					let p12Base64 = queryValue("p12"),
+					let provisionBase64 = queryValue("mobileprovision"),
+					let passwordBase64 = queryValue("password"),
+					let passwordData = Data(base64Encoded: passwordBase64),
+					let password = String(data: passwordData, encoding: .utf8)
+				else {
+					return false
+				}
+				
+				let generator = UINotificationFeedbackGenerator()
+				generator.prepare()
+				
+				guard
+					let p12URL = FileManager.default.decodeAndWrite(base64: p12Base64, pathComponent: ".p12"),
+					let provisionURL = FileManager.default.decodeAndWrite(base64: provisionBase64, pathComponent: ".mobileprovision"),
+					FR.checkPasswordForCertificate(for: p12URL, with: password, using: provisionURL)
+				else {
+					generator.notificationOccurred(.error)
+					return false
+				}
+				
+				FR.handleCertificateFiles(
+					p12URL: p12URL,
+					provisionURL: provisionURL,
+					p12Password: password
+				) { error in
+					if let error = error {
+						UIAlertController.showAlertWithOk(title: .localized("Error"), message: error.localizedDescription)
+					} else {
+						generator.notificationOccurred(.success)
+					}
+				}
+				
+				return true
+			}
+			/// feather://source/<url>
+			if let fullPath = url.validatedScheme(after: "/source/") {
+				FR.handleSource(fullPath) { }
+			}
+			/// feather://install/<url.ipa>
+			if
+				let fullPath = url.validatedScheme(after: "/install/"),
+				let downloadURL = URL(string: fullPath)
+			{
+				_ = DownloadManager.shared.startDownload(from: downloadURL)
+			}
+		} else {
+			if url.pathExtension == "ipa" || url.pathExtension == "tipa" {
+				if FileManager.default.isFileFromFileProvider(at: url) {
+					guard url.startAccessingSecurityScopedResource() else { return false }
+					FR.handlePackageFile(url) { _ in }
+				} else {
+					FR.handlePackageFile(url) { _ in }
+				}
+				
+				return true
+			}
+		}
+		
+		return false
+	}
+
+}
